+++ inst/doc/introduction-to-gggenes.R
@@ -78,3 +78,19 @@ ggplot(subset(example_genes, molecule == "Genome4" & gene == "genA"),
     min.size = 0
   )
 
+## -----------------------------------------------------------------------------
+ggplot(example_genes, aes(xmin = start, xmax = end, y = molecule, fill = gene)) +
+  geom_feature(
+    data = example_features,
+    aes(x = position, y = molecule, forward = forward)
+  ) +
+  geom_feature_label(
+    data = example_features,
+    aes(x = position, y = molecule, label = name, forward = forward)
+  ) +
+  geom_gene_arrow() +
+  geom_blank(data = example_dummies) +
+  facet_wrap(~ molecule, scales = "free", ncol = 1) +
+  scale_fill_brewer(palette = "Set3") +
+  theme_genes()
+